+++ cantera18/branches/scons/Cantera/src/thermo/Elements.h
@@ -97,7 +97,7 @@ namespace Cantera {
     /*!
      *  @param m element index
      */
-    doublereal atomicWeight(int m) const { return m_atomicWeights[m]; }
+    doublereal atomicWeight(size_t m) const { return m_atomicWeights[m]; }
 
     /// Atomic number of element m.
     /*!
@@ -112,7 +112,7 @@ namespace Cantera {
      *
      *  @param m  Element index
      */
-    doublereal entropyElement298(int m) const;
+    doublereal entropyElement298(size_t m) const;
 
     /// vector of element atomic weights
     const vector_fp& atomicWeights() const { return m_atomicWeights; }
@@ -123,7 +123,7 @@ namespace Cantera {
      *  @return 
      *    \c int: The number of elements in the object.
      */
-    int nElements() const { return m_mm; }
+    size_t nElements() const { return m_mm; }
 
     //! Function that returns the index of an element.
     /*!
@@ -135,7 +135,7 @@ namespace Cantera {
      *
      * @param name String containing the index.
      */
-    int elementIndex(std::string name) const;
+    size_t elementIndex(std::string name) const;
       
     //! Name of the element with index \c m. 
     /*!
